+++ app/Http/Controllers/Backend/UserController.php
@@ -9,32 +9,39 @@ use Illuminate\Support\Arr;
 use App\Http\Requests\UserRequest;
 use Spatie\Permission\Models\Role;
 use App\Services\MediaStoreService;
+use Illuminate\Contracts\View\View;
 use App\Http\Controllers\Controller;
-use Spatie\Permission\Models\Permission;
+use Illuminate\Http\RedirectResponse;
+use App\Services\ChunkPermissionService;
 
 class UserController extends Controller
 {
-    public function index() {
+    public function index(): View  {
         $users = User::withCount('permissions')->with('roles', 'media')->paginate(10);
 
         return view('backend.users.index', compact('users'));
     }
 
-    public function create() {
-        $roles = Role::all();
+    public function create(): View  {
+        $roles = Role::where('id', '>', 1)->get();
         $userRoles = [];
-        $permissions = Permission::all();
+        $permissions = (new ChunkPermissionService())->permission;
         $userPermissions = [];
 
         return view('backend.users.create', compact('roles', 'userRoles', 'permissions', 'userPermissions'));
     }
 
-    public function store(UserRequest $request, User $user, MediaStoreService $mediaService) {
+    public function store(UserRequest $request, User $user, MediaStoreService $mediaService): RedirectResponse {
         $validated = $request->validated();
         $user->create($validated);
 
-        // store rols to user
-        $role = $request->input('role');
+        // store roles to user
+        $role = collect($request->input('role'))
+            ->filter(function ($value, $key) {
+                return $value >= 2; // SuperAdmin remove
+            })->when($user->id == 1, function ($collection) {
+                return $collection->push(1); // SuperAdmin add
+            })->toArray();
         $user->roles()->sync($role);
 
         // store permissions to user
@@ -42,40 +49,57 @@ class UserController extends Controller
         $user->permissions()->sync($permissions);
 
         if ($request->hasFile('photo_avatar')) {
-            $mediaService->storeMediaOneFile($user, 'avatar', 'photo_avatar');
+            $mediaService->storeMediaOneFile($user, $user->collectionName, 'photo_avatar');
         }
 
-        toastr()->success(__('app.user.store'));
-        return redirect()->route('users.index');
+        toastr()->success(__('app.user.store', ['name'=> $user->name]));
+        return to_route('users.index');
     }
 
-    public function show($id) {
-        $user = User::whereId($id)->withCount('permissions')->with('roles', 'media')->firstOrFail();
+    public function show(User $user): View  {
+        $user->with('roles', 'media')->withCount('permissions');
 
-        return view('backend.users.show', compact( 'user' ) );
+        return view('backend.users.show', compact('user') );
     }
 
-    public function edit(User $user) {
-        $roles = Role::all();
+    public function edit(User $user): View  {
+        if ($user->id == 1 AND auth()->user()->id != 1) {
+            toastr()->error(__('app.user.update-error', ['name'=> $user->name]));
+            return to_route('users.index');
+        }
+        $roles = Role::where('id', '>', 1)->get();
         $userRoles = $user->roles->pluck('id')->toArray();
-        $permissions = Permission::all();
+        $permissions = (new ChunkPermissionService())->permission;
         $userPermissions = $user->permissions->pluck('id')->toArray();
 
         return view('backend.users.edit', compact('user', 'roles', 'userRoles', 'permissions', 'userPermissions'));
     }
 
-    public function update(UserRequest $request, User $user, MediaStoreService $mediaService) {
+    public function update(UserRequest $request, User $user, MediaStoreService $mediaService): RedirectResponse {
         $validated = $request->validated();
 
         // if no password is entered, it is removed from the request
-        if( ! $request->filled('password') ) {
+        if (! $request->filled('password')) {
             $validated = Arr::except($validated, ['password']);
+        }
+
+        // if user change self
+        if ($user->id == auth()->user()->id) {
+            $validated = Arr::except($validated, ['active']);
+            toastr()->warning(__('app.user.update-self'));
+        } else {
+            toastr()->success(__('app.user.update', ['name'=> $user->name]));
         }
 
         $user->update($validated);
 
-        // store rols to user
-        $role = $request->input('role');
+        // store roles to user
+        $role = collect($request->input('role'))
+            ->filter(function ($value, $key) {
+                return $value >= 2; // SuperAdmin remove
+            })->when($user->id == 1, function ($collection) {
+                return $collection->push(1);  // SuperAdmin add
+            })->toArray();
         $user->roles()->sync($role);
 
         // store permissions to user
@@ -83,17 +107,23 @@ class UserController extends Controller
         $user->permissions()->sync($permissions);
 
         if ($request->hasFile('photo_avatar')) {
-            $mediaService->storeMediaOneFile($user, 'avatar', 'photo_avatar');
+            $mediaService->storeMediaOneFile($user, $user->collectionName, 'photo_avatar');
         }
 
-        toastr()->success(__('app.user.update'));
-        return redirect()->route('users.index');
+        return to_route('users.index');
     }
 
-    public function destroy(User $user) {
-        $user->delete();
+    public function destroy(User $user): RedirectResponse {
+        if ($user->id == 1) {
+            toastr()->error(__('app.user.delete-error', ['name'=> $user->name]));
+        } elseif ($user->id == auth()->user()->id) {
+            toastr()->error(__('app.user.delete-self'));
+        } else {
+            $user->delete();
+            $user->clearMediaCollection($user->collectionName);
+            toastr()->success(__('app.user.delete', ['name'=> $user->name]));
+        }
 
-        toastr()->success(__('app.user.delete'));
-        return redirect()->route('users.index');
+        return to_route('users.index');
     }
 }
